# Application/Frigorino.Web/ClientApp/src/lib/api/models/CreateInventoryItemRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateInventoryItemRequest = {
    text?: string | null;
    quantity?: string | null;
    expiryDate?: string | null;
};