# Application/Frigorino.Web/ClientApp/src/lib/api/models/InventoryItemDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type InventoryItemDto = {
    id?: number;
    inventoryId?: number;
    text?: string | null;
    quantity?: string | null;
    expiryDate?: string | null;
    sortOrder?: number;
    createdAt?: string;
    updatedAt?: string;
    isExpiring?: boolean;
};