# Application/Frigorino.Web/ClientApp/src/lib/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Frigorino.Web",
        "version": "1.0"
    },
    "paths": {
...
        "/api/inventory/{inventoryId}/InventoryItems": {
            "get": {
                "tags": ["InventoryItems"],
                "parameters": [
                    {
                        "name": "inventoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InventoryItemDto"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InventoryItemDto"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InventoryItemDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["InventoryItems"],
                "parameters": [
                    {
                        "name": "inventoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateInventoryItemRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateInventoryItemRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateInventoryItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryItemDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryItemDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryItemDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/inventory/{inventoryId}/InventoryItems/{inventoryItemId}": {
            "get": {
                "tags": ["InventoryItems"],
                "parameters": [
                    {
                        "name": "inventoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "inventoryItemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryItemDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryItemDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryItemDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["InventoryItems"],
                "parameters": [
                    {
                        "name": "inventoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "inventoryItemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateInventoryItemRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateInventoryItemRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateInventoryItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryItemDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryItemDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryItemDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["InventoryItems"],
                "parameters": [
                    {
                        "name": "inventoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "inventoryItemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/Items/{itemId}": {
            "get": {
                "tags": ["Items"],
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListItemDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListItemDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListItemDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["Items"],
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateListItemRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateListItemRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateListItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListItemDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListItemDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListItemDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Items"],
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateInventoryItemRequest": {
                "type": "object",
                "properties": {
                    "text": {
                        "type": "string",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "string",
                        "nullable": true
                    },
                    "expiryDate": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
            "InventoryItemDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "inventoryId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "text": {
                        "type": "string",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "string",
                        "nullable": true
                    },
                    "expiryDate": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "sortOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "isExpiring": {
                        "type": "boolean"
                    }
                },
                "additionalProperties": false
            },
...
            "ListItemDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "listId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "text": {
                        "type": "string",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "sortOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "additionalProperties": false
            },
...
            "UpdateInventoryItemRequest": {
                "type": "object",
                "properties": {
                    "text": {
                        "type": "string",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "string",
                        "nullable": true
                    },
                    "expiryDate": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
            "UpdateListItemRequest": {
                "type": "object",
                "properties": {
                    "text": {
                        "type": "string",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "boolean",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
        }
    }
}